refactor(harness): move the OpenAI harness into adk/_modules + export at facade
Promote the OpenAI Turn and sync tap to the canonical adk/_modules layout
(_openai_turn.py, _openai_sync.py), matching every other harness. Leave
back-compat shims at providers/_modules/{openai_turn,sync_provider}.py so
existing import paths keep working; the larger Temporal/MCP provider stays
under adk.providers.

Export OpenAITurn, openai_usage_to_turn_usage and convert_openai_to_agentex_events
from agentex.lib.adk so the OpenAI harness is a first-class facade citizen like
ClaudeCodeTurn / CodexTurn (fixes Greptile follow-up: templates can import from
the package facade instead of a private module path).
